Create separate eprbuild and buildtypes modules
