+++ netqasm/sdk/builder.py
@@ -7,20 +7,8 @@ Python application script code into NetQASM subroutines.
 from __future__ import annotations
 
 from contextlib import contextmanager
-from dataclasses import dataclass
 from itertools import count
-from typing import (
-    TYPE_CHECKING,
-    Callable,
-    Dict,
-    Iterator,
-    List,
-    Optional,
-    Set,
-    Tuple,
-    Type,
-    Union,
-)
+from typing import Dict, Iterator, List, Optional, Set, Tuple, Type, Union
 
 from netqasm import NETQASM_VERSION
 from netqasm.backend.network_stack import OK_FIELDS_K, OK_FIELDS_M
@@ -42,217 +30,29 @@ from netqasm.lang.ir import (
 )
 from netqasm.lang.parsing.text import assemble_subroutine, parse_register
 from netqasm.lang.subroutine import Subroutine
-from netqasm.qlink_compat import (
-    BellState,
-    EPRRole,
-    EPRType,
-    LinkLayerOKTypeK,
-    LinkLayerOKTypeM,
-    LinkLayerOKTypeR,
-    RandomBasis,
-    TimeUnit,
+from netqasm.qlink_compat import EPRRole, EPRType, LinkLayerOKTypeK
+from netqasm.sdk.buildtypes import (
+    T_BranchRoutine,
+    T_LinkLayerOkList,
+    T_LoopRoutine,
+    T_PostRoutine,
 )
 from netqasm.sdk.compiling import NVSubroutineCompiler, SubroutineCompiler
 from netqasm.sdk.config import LogConfig
+from netqasm.sdk.eprbuild import (
+    EntRequestParams,
+    EprKeepResult,
+    EprMeasureResult,
+    deserialize_epr_keep_results,
+    deserialize_epr_measure_results,
+    serialize_request,
+)
 from netqasm.sdk.futures import Array, Future, RegFuture, T_CValue
 from netqasm.sdk.memmgr import MemoryManager
 from netqasm.sdk.qubit import FutureQubit, Qubit
 from netqasm.sdk.toolbox import get_angle_spec_from_float
 from netqasm.typedefs import T_Cmd
 from netqasm.util.log import LineTracker
-
-T_LinkLayerOkList = Union[
-    List[LinkLayerOKTypeK], List[LinkLayerOKTypeM], List[LinkLayerOKTypeR]
-]
-T_PostRoutine = Callable[
-    ["Builder", Union[FutureQubit, List[Future]], operand.Register], None
-]
-T_BranchRoutine = Callable[["connection.BaseNetQASMConnection"], None]
-T_LoopRoutine = Callable[["connection.BaseNetQASMConnection"], None]
-
-if TYPE_CHECKING:
-    from . import connection
-
-
-@dataclass
-class EntRequestParams:
-    remote_node_id: int
-    epr_socket_id: int
-    number: int
-    post_routine: Optional[T_PostRoutine]
-    sequential: bool
-    time_unit: TimeUnit = TimeUnit.MICRO_SECONDS
-    max_time: int = 0
-    random_basis_local: Optional[RandomBasis] = None
-    random_basis_remote: Optional[RandomBasis] = None
-    rotations_local: Tuple[int, int, int] = (0, 0, 0)
-    rotations_remote: Tuple[int, int, int] = (0, 0, 0)
-
-
-# Length of NetQASM array for serialized Create Requests.
-SER_CREATE_REQUEST_LEN = 20
-
-# Indices of Create Request arguments in serialized NetQASM array
-SER_CREATE_IDX_TYPE = 0
-SER_CREATE_IDX_NUMBER = 1
-SER_CREATE_IDX_RANDOM_BASIS_LOCAL = 2
-SER_CREATE_IDX_RANDOM_BASIS_REMOTE = 3
-SER_CREATE_IDX_MINIMUM_FIDELITY = 4
-SER_CREATE_IDX_TIME_UNIT = 5
-SER_CREATE_IDX_MAX_TIME = 6
-SER_CREATE_IDX_PRIORITY = 7
-SER_CREATE_IDX_ATOMIC = 8
-SER_CREATE_IDX_CONSECUTIVE = 9
-SER_CREATE_IDX_PROBABILITY_DIST_LOCAL1 = 10
-SER_CREATE_IDX_PROBABLIITY_DIST_LOCAL2 = 11
-SER_CREATE_IDX_PROBABILITY_DIST_REMOTE1 = 12
-SER_CREATE_IDX_PROBABLIITY_DIST_REMOTE2 = 13
-SER_CREATE_IDX_ROTATION_X_LOCAL1 = 14
-SER_CREATE_IDX_ROTATION_Y_LOCAL = 15
-SER_CREATE_IDX_ROTATION_X_LOCAL2 = 16
-SER_CREATE_IDX_ROTATION_X_REMOTE1 = 17
-SER_CREATE_IDX_ROTATION_Y_REMOTE = 18
-SER_CREATE_IDX_ROTATION_X_REMOTE2 = 19
-
-# Length of NetQASM array for EPR Keep results.
-SER_RESULT_KEEP_LEN = 10
-
-# Indices of EPR Keep results in serialized NetQASM array
-SER_RESPONSE_KEEP_IDX_TYPE = 0
-SER_RESPONSE_KEEP_IDX_CREATE_ID = 1
-SER_RESPONSE_KEEP_IDX_LOGICAL_QUBIT_ID = 2
-SER_RESPONSE_KEEP_IDX_DIRECTONIALITY_FLAG = 3
-SER_RESPONSE_KEEP_IDX_SEQUENCE_NUMBER = 4
-SER_RESPONSE_KEEP_IDX_PURPOSE_ID = 5
-SER_RESPONSE_KEEP_IDX_REMOTE_NODE_ID = 6
-SER_RESPONSE_KEEP_IDX_GOODNESS = 7
-SER_RESPONSE_KEEP_IDX_GOODNESS_TIME = 8
-SER_RESPONSE_KEEP_IDX_BELL_STATE = 9
-
-# Length of NetQASM array for EPR Measure results.
-SER_RESULT_MEASURE_LEN = 10
-
-# Indices of EPR Measure results in serialized NetQASM array
-SER_RESPONSE_MEASURE_IDX_TYPE = 0
-SER_RESPONSE_MEASURE_IDX_CREATE_ID = 1
-SER_RESPONSE_MEASURE_IDX_MEASUREMENT_OUTCOME = 2
-SER_RESPONSE_MEASURE_IDX_MEASUREMENT_BASIS = 3
-SER_RESPONSE_MEASURE_IDX_DIRECTONIALITY_FLAG = 4
-SER_RESPONSE_MEASURE_IDX_SEQUENCE_NUMBER = 5
-SER_RESPONSE_MEASURE_IDX_PURPOSE_ID = 6
-SER_RESPONSE_MEASURE_IDX_REMOTE_NODE_ID = 7
-SER_RESPONSE_MEASURE_IDX_GOODNESS = 8
-SER_RESPONSE_MEASURE_IDX_BELL_STATE = 9
-
-
-def serialize_request(tp: EPRType, params: EntRequestParams) -> List[Optional[int]]:
-    array: List[Optional[int]] = [None for i in range(SER_CREATE_REQUEST_LEN)]
-
-    array[SER_CREATE_IDX_TYPE] = tp.value
-    array[SER_CREATE_IDX_NUMBER] = params.number
-
-    # Only when max_time is 0, explicitly initialize the relavant array elements.
-    # If it is max_time 0, these array element will be None.
-    if params.max_time != 0:
-        array[SER_CREATE_IDX_TIME_UNIT] = params.time_unit.value
-        array[SER_CREATE_IDX_MAX_TIME] = params.max_time
-
-    if tp == EPRType.M or tp == EPRType.R:
-        # Only write when non-zero.
-        if params.rotations_local != (0, 0, 0):
-            array[SER_CREATE_IDX_ROTATION_X_LOCAL1] = params.rotations_local[0]
-            array[SER_CREATE_IDX_ROTATION_Y_LOCAL] = params.rotations_local[1]
-            array[SER_CREATE_IDX_ROTATION_X_LOCAL2] = params.rotations_local[2]
-        if params.rotations_remote != (0, 0, 0):
-            array[SER_CREATE_IDX_ROTATION_X_REMOTE1] = params.rotations_remote[0]
-            array[SER_CREATE_IDX_ROTATION_Y_REMOTE] = params.rotations_remote[1]
-            array[SER_CREATE_IDX_ROTATION_X_REMOTE2] = params.rotations_remote[2]
-        if params.random_basis_local:
-            array[SER_CREATE_IDX_RANDOM_BASIS_LOCAL] = params.random_basis_local.value
-        if params.random_basis_remote:
-            array[SER_CREATE_IDX_RANDOM_BASIS_REMOTE] = params.random_basis_remote.value
-
-    return array
-
-
-def deserialize_epr_keep_results(num_pairs: int, array: Array) -> List[EprKeepResult]:
-    assert len(array) == num_pairs * SER_RESULT_KEEP_LEN
-    results: List[EprKeepResult] = []
-    for i in range(num_pairs):
-        base = i * SER_RESULT_KEEP_LEN
-        results.append(
-            EprKeepResult(
-                qubit_id=array.get_future_index(
-                    base + SER_RESPONSE_KEEP_IDX_LOGICAL_QUBIT_ID
-                ),
-                remote_node_id=array.get_future_index(
-                    base + SER_RESPONSE_KEEP_IDX_REMOTE_NODE_ID
-                ),
-                generation_duration=array.get_future_index(
-                    base + SER_RESPONSE_KEEP_IDX_GOODNESS
-                ),
-                raw_bell_state=array.get_future_index(
-                    base + SER_RESPONSE_KEEP_IDX_BELL_STATE
-                ),
-            )
-        )
-    return results
-
-
-def deserialize_epr_measure_results(
-    num_pairs: int, array: Array
-) -> List[EprMeasureResult]:
-    assert len(array) == num_pairs * SER_RESULT_MEASURE_LEN
-    results: List[EprMeasureResult] = []
-    for i in range(num_pairs):
-        base = i * SER_RESULT_MEASURE_LEN
-        results.append(
-            EprMeasureResult(
-                measurement_outcome=array.get_future_index(
-                    base + SER_RESPONSE_MEASURE_IDX_MEASUREMENT_OUTCOME
-                ),
-                measurement_basis=array.get_future_index(
-                    base + SER_RESPONSE_MEASURE_IDX_MEASUREMENT_BASIS
-                ),
-                remote_node_id=array.get_future_index(
-                    base + SER_RESPONSE_MEASURE_IDX_REMOTE_NODE_ID
-                ),
-                generation_duration=array.get_future_index(
-                    base + SER_RESPONSE_MEASURE_IDX_GOODNESS
-                ),
-                raw_bell_state=array.get_future_index(
-                    base + SER_RESPONSE_MEASURE_IDX_BELL_STATE
-                ),
-            )
-        )
-    return results
-
-
-@dataclass
-class EprKeepResult:
-    qubit_id: Future
-    remote_node_id: Future
-    generation_duration: Future
-    raw_bell_state: Future
-
-    @property
-    def bell_state(self) -> BellState:
-        assert self.raw_bell_state.value is not None
-        return BellState(self.raw_bell_state.value)
-
-
-@dataclass
-class EprMeasureResult:
-    measurement_outcome: Future
-    measurement_basis: Future
-    remote_node_id: Future
-    generation_duration: Future
-    raw_bell_state: Future
-
-    @property
-    def bell_state(self) -> BellState:
-        assert self.raw_bell_state.value is not None
-        return BellState(self.raw_bell_state.value)
 
 
 class LabelManager:
